VPLAY-12576  Add precision to AAMP totalPerformRequest HttpRequestEnd
Reason for change: To avoid rounding of AAMP segment's downloadTimeMS
Risks: Low
Priority: P1

+++ drm/AampDRMLicManager.h
@@ -191,7 +191,7 @@ public:
 	 * @param[in] eventHandle - DRM Metadata event handle
 	 * @param[in] respData - download response data
 	 */
-	void UpdateLicenseMetrics(DrmRequestType requestType, int32_t statusCode, std::string licenseRequestUrl, long long downloadTimeMS, DrmMetaDataEventPtr eventHandle, DownloadResponsePtr respData );
+	void UpdateLicenseMetrics(DrmRequestType requestType, int32_t statusCode, std::string licenseRequestUrl, double downloadTimeMS, DrmMetaDataEventPtr eventHandle, DownloadResponsePtr respData );
 	/**
 	 * @fn ContentProtectionDataUpdate
 	 */
